Move timing instrumentation related code into scalyr_agent.instrumentation package.
